refactor(helper-image): removing indirection and putting the implementation in index.js
